refactor: 클라이언트 컴포넌트 폴더 경로 수정
