+++ src/components/mypage/album-analysis-page.tsx
